+++ emcc.py
@@ -27,7 +27,6 @@ import logging
 import os
 import re
 import shlex
-import shutil
 import sys
 import time
 import tarfile
@@ -78,7 +77,6 @@ EXTRA_INCOMING_JS_API = [
 
 SIMD_INTEL_FEATURE_TOWER = ['-msse', '-msse2', '-msse3', '-mssse3', '-msse4.1', '-msse4.2', '-msse4', '-mavx']
 SIMD_NEON_FLAGS = ['-mfpu=neon']
-COMPILE_ONLY_FLAGS = {'--default-obj-ext'}
 LINK_ONLY_FLAGS = {
     '--bind', '--closure', '--cpuprofiler', '--embed-file',
     '--emit-symbol-map', '--emrun', '--exclude-file', '--extern-post-js',
@@ -149,7 +147,6 @@ class EmccOptions:
     self.memory_init_file = None
     self.use_preload_cache = False
     self.use_preload_plugins = False
-    self.default_object_extension = '.o'
     self.valid_abspaths = []
     # Specifies the line ending format to use for all generated text files.
     # Defaults to using the native EOL on each platform (\r\n on Windows, \n on
@@ -811,23 +808,6 @@ def phase_setup(options, state, newargs):
         diagnostics.warning(
             'unused-command-line-argument',
             "linker flag ignored during compilation: '%s'" % arg)
-    if state.has_dash_c:
-      if '-emit-llvm' in newargs:
-        options.default_object_extension = '.bc'
-    elif state.has_dash_S:
-      if '-emit-llvm' in newargs:
-        options.default_object_extension = '.ll'
-      else:
-        options.default_object_extension = '.s'
-    elif '-M' in newargs or '-MM' in newargs:
-      options.default_object_extension = '.mout' # not bitcode, not js; but just dependency rule of the input file
-
-  else:
-    for arg in state.orig_args:
-      if any(arg.startswith(f) for f in COMPILE_ONLY_FLAGS):
-        diagnostics.warning(
-            'unused-command-line-argument',
-            "compiler flag ignored during linking: '%s'" % arg)
 
   if settings.MAIN_MODULE or settings.SIDE_MODULE:
     settings.RELOCATABLE = 1
@@ -1006,20 +986,7 @@ def phase_compile_inputs(options, state, newargs, input_files):
       cmd += ['-o', options.output_file]
       if get_file_suffix(options.output_file) == '.bc' and not settings.LTO and '-emit-llvm' not in state.orig_args:
         diagnostics.warning('emcc', '.bc output file suffix used without -flto or -emit-llvm.  Consider using .o extension since emcc will output an object file, not a bitcode file')
-    ext = get_clang_output_extension(state)
-    if not options.output_file and options.default_object_extension != ext:
-      # If we are using a non-standard output file extention we cannot use
-      # exec_process here since we need to rename the files
-      # after clang runs (since clang does not support --default-obj-ext)
-      # TODO: Remove '--default-obj-ext' to reduce this complexity
-      shared.check_call(cmd)
-      for i in inputs:
-        output = unsuffixed_basename(i) + ext
-        new_output = unsuffixed_basename(i) + options.default_object_extension
-        shutil.move(output, new_output)
-      sys.exit(0)
-    else:
-      shared.exec_process(cmd)
+    shared.exec_process(cmd)
 
   # In COMPILE_AND_LINK we need to compile source files too, but we also need to
   # filter out the link flags
@@ -1039,7 +1006,7 @@ def phase_compile_inputs(options, state, newargs, input_files):
     return unsuffixed(name) + '_' + seen_names[name] + shared.suffix(name)
 
   def get_object_filename(input_file):
-    return in_temp(unsuffixed(uniquename(input_file)) + options.default_object_extension)
+    return in_temp(shared.replace_suffix(uniquename(input_file), '.o'))
 
   def compile_source_file(i, input_file):
     logger.debug(f'compiling source file: {input_file}')
@@ -1384,9 +1351,7 @@ def parse_args(newargs):
     elif arg == '-fignore-exceptions':
       settings.DISABLE_EXCEPTION_CATCHING = 1
     elif check_arg('--default-obj-ext'):
-      options.default_object_extension = consume_arg()
-      if not options.default_object_extension.startswith('.'):
-        options.default_object_extension = '.' + options.default_object_extension
+      exit_with_error('--default-obj-ext is no longer supported by emcc')
     elif arg.startswith('-fsanitize=cfi'):
       exit_with_error('emscripten does not currently support -fsanitize=cfi')
     elif check_arg('--output_eol'):